+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 1.7.0+enfold5 (unreleased)
 --------------------------
 
-- Nothing changed yet.
+- Upload pattern uses the baseUrl to compute the upload URL, so this should
+  always be the site root and not the current context
+  [frapell]
 
 
 1.7.0+enfold4 (2015-06-16)